ci: Drop explicit "connection:" for provisioning
This forces having to have a system libvirt instance, but that is
incompatible with running tests inside of containers/toolbox. It also
seems unnecessary: session libvirt networking and the tests work just
fine.

Drop the explicit setting and let tmt/testcloud pick its internal
default (which is `qemu:///session`). This wasn't relevant for running
on Testing Farm anyway.

Signed-off-by: Martin Pitt <mpitt@redhat.com>