ci: Remove tft trigger capability for contributors
The tft jobs run inside the RH internal network, and thus have access to
a lot of sensitive data. It is very easy to gain "contributor" status in
GitHub, you just need to land a single commit (like a typo fix or small
doc update), and it is hard to get a list of/control the set of
contributors.

Thus restrict the trigger capability to org members and collaborators,
which are an explicit list and much easier to control.