ci: Run QEMU tox integration tests in GitHub workflow
GitHub workflows recently grew the ability to set up nested QEMU kvm,
but without much fanfare (see [1]). It just needs to create the
`/dev/kvm` device node, but the host has it enabled and the kernel
supports it.

With that we can run the QEMU tox tests. They are not covered anywhere
else, and currently have some bugs, so let's cover them in CI and keep
them green.

[1] https://github.com/orgs/community/discussions/8305