ci: ansible-lint needs collection path; tox-lsr 3.5.1
Without this - what happens is that the new ansible-lint will install
the collection dependencies in the requirements file in the first
place in ANSIBLE_COLLECTIONS_PATH which will be used instead of
the current code that has been converted to collection format.  When
this happens, any modules/plugins that your PR has added will not be
found.

With this fix, ansible-lint will not overwrite the existing code
which has been converted to collection format.

bump tox-lsr version to 3.5.1